refaktor schema syncer
